// cloudreve-official-frontend/src/api/dashboard.ts
export interface GroupEnt extends CommonMixin {
  name: string;
  max_storage?: number;
  speed_limit?: number;
  permissions?: string;
  edges: {
    storage_policies: StoragePolicy[];
  };
  total_users?: number;
  settings?: GroupSetting;
}
// ...
export interface CommonMixin {
  id: number;
  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}

export interface StoragePolicy extends CommonMixin {
  name: string;
  type: PolicyType;
  server?: string;
  bucket_name?: string;
  is_private?: boolean;
  access_key?: string;
  secret_key?: string;
  max_size?: number;
  auto_rename?: boolean;
  dir_name_rule?: string;
  file_name_rule?: string;
  settings?: PolicySetting;
  node_id?: number;
  edges: {
    users?: User[];
    groups?: GroupEnt[];
    node?: Node;
    policy_load_balance?: PolicyLoadBalance[];
  };
  entities_count?: number;
  entities_size?: number;
  hash_id?: string;
}
// ...
export interface User extends CommonMixin {
  email: string;
  nick: string;
  password?: string;
  settings?: UserSetting;
  status?: UserStatus;
  storage?: number;
  avatar?: string;
  credit?: number;
  group_expires?: string;
  notify_date?: string;
  group_users?: number;
  previous_group?: number;
  unmanaged_email?: boolean;
  edges: {
    group?: GroupEnt;
    storage_policy?: StoragePolicy;
    openid?: OpenID[];
    passkey?: Passkey[];
  };

  hash_id?: string;
  two_fa_enabled?: boolean;
  capacity?: Capacity;
}

export interface OpenID extends CommonMixin {
  provider?: number;
}

export interface Passkey extends CommonMixin {
  name?: string;
}

export enum UserStatus {
  active = "active",
  inactive = "inactive",
  manual_banned = "manual_banned",
  sys_banned = "sys_banned",
}

export interface UserSetting {
  profile_off?: boolean;
  preferred_policy?: number;
  preferred_theme?: string;
  version_retention?: boolean;
  version_retention_ext?: string[];
  version_retention_max?: number;
  pined?: PinedFile[];
  language?: string;
}
// ...
export interface PaidProductProps {
  share_link_id?: number;
  points_gain?: number;
  group_id?: number;
  storage_size?: number;
  duration?: number;
  sku_id?: string;
}

export interface Payment extends CommonMixin {
  trade_no?: string;
  product_type?: number;
  product_props?: PaidProductProps;
  provider_id?: string;
  name?: string;
  price_unit?: number;
  price_mark?: string;
  price_id?: string;
  provider_reference_id?: string;
  price_one_unit?: number;
  qyt?: number;
  status?: PaymentStatus;
  anonymous_email?: string;
  creator?: number;
  edges: {
    user?: User;
  };

  user_hash_id?: string;
}